[ai-assisted] refactor(objecttype): store contract cleanup and yaml rebind semantics (#163)
Why:
- ObjectTypeStore 인터페이스가 db.jdbc.model 패키지의 Row 클래스를 계약으로 노출해
  JDBC 구현 세부사항이 store 계약에 누출되어 있었다.
- YAML 모드에서 objectTypeRegistry, objectPolicyResolver 빈이 각각 loader.load()를
  호출해 같은 YAML을 두 번 파싱했다.
- YamlObjectRebindService의 rebind 의미(cache 무효화 vs. reload)가 코드와 문서
  어디에도 명시되지 않아 운영 의미가 불명확했다.
- objectRebindService 빈 등록 로그가 YamlObjectPolicyResolver를 잘못 참조했다.
- AttachmentEndpointAutoConfiguration에 @ComponentScan과 @Bean이 동시에 존재해
  등록 경로가 중복되어 있었다.

What:
- ObjectTypeRow, ObjectTypePolicyRow를 db.jdbc.model → db.model로 이동해
  store 계약에서 jdbc 패키지 의존을 제거했다.
- yamlObjectTypeLoaderResult @Bean을 추가해 YAML 파싱을 한 번만 수행하고
  registry, policy resolver가 결과를 공유하도록 변경했다.
- YamlObjectRebindService에 Javadoc을 추가해 rebind = 캐시 무효화만 의미하며
  YAML 재적재는 발생하지 않음을 명시했다.
- objectRebindService 빈 로그를 YamlObjectRebindService로 수정했다.
- AttachmentEndpointAutoConfiguration에서 @ComponentScan을 제거했다.

Validation:
- ./gradlew :studio-platform-objecttype:compileJava → BUILD SUCCESSFUL
- ./gradlew :starter:studio-platform-starter-objecttype:compileJava → BUILD SUCCESSFUL
- ./gradlew :starter:studio-application-starter-attachment:compileJava → BUILD SUCCESSFUL
- ./gradlew :studio-platform-objecttype:test → BUILD SUCCESSFUL
- ./gradlew :starter:studio-platform-starter-objecttype:test → BUILD SUCCESSFUL
