+++ studio-platform-objecttype/src/main/java/studio/one/platform/objecttype/db/jdbc/ObjectTypeJdbcRepository.java
@@ -16,8 +16,8 @@ import org.springframework.util.StringUtils;
 
 import studio.one.platform.data.sqlquery.annotation.SqlStatement;
 import studio.one.platform.objecttype.db.ObjectTypeStore;
-import studio.one.platform.objecttype.db.jdbc.model.ObjectTypePolicyRow;
-import studio.one.platform.objecttype.db.jdbc.model.ObjectTypeRow;
+import studio.one.platform.objecttype.db.model.ObjectTypePolicyRow;
+import studio.one.platform.objecttype.db.model.ObjectTypeRow;
 
 public class ObjectTypeJdbcRepository implements ObjectTypeStore {
