+++ studio-platform-objecttype/src/main/java/studio/one/platform/objecttype/db/jdbc/model/JdbcObjectTypeMetadata.java
@@ -4,6 +4,7 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 
+import studio.one.platform.objecttype.db.model.ObjectTypeRow;
 import studio.one.platform.objecttype.model.ObjectTypeMetadata;
 
 public class JdbcObjectTypeMetadata implements ObjectTypeMetadata {
