+++ studio-platform-objecttype/src/main/java/studio/one/platform/objecttype/db/model/ObjectTypeRow.java
@@ -1,4 +1,4 @@
-package studio.one.platform.objecttype.db.jdbc.model;
+package studio.one.platform.objecttype.db.model;
 
 import java.time.Instant;
 
@@ -7,12 +7,13 @@ import lombok.Setter;
 
 @Getter
 @Setter
-public class ObjectTypePolicyRow {
+public class ObjectTypeRow {
     private int objectType;
-    private Integer maxFileMb;
-    private String allowedExt;
-    private String allowedMime;
-    private String policyJson;
+    private String code;
+    private String name;
+    private String domain;
+    private String status;
+    private String description;
     private String createdBy;
     private long createdById;
     private Instant createdAt;
